Creating tests@

+++ src/main/java/br/com/fiap/FarmaNear_Register/controller/dto/DrugstoreDto.java
@@ -1,8 +1,8 @@
 package br.com.fiap.FarmaNear_Register.controller.dto;
 
-public record DrugstoreDto(Long id, Integer cnpj, String name, String email, String phone, AddressDto address) {
+public record DrugstoreDto(Long id, Long cnpj, String name, String email, String phone, AddressDto address) {
 
-    public DrugstoreDto(long id, Integer cnpj, String name, String email, String phone) {
+    public DrugstoreDto(Long id, Long cnpj, String name, String email, String phone) {
         this(id, cnpj, name, email, phone, null);
     }
 }
