+++ src/main/java/br/com/fiap/FarmaNear_Register/entities/drugstore/Drugstore.java
@@ -8,14 +8,14 @@ import br.com.fiap.FarmaNear_Register.infra.repository.drugstore.DrugstoreEntity
 import java.util.Set;
 
 public class Drugstore implements DrugstoreEntityInterface {
-    private final Integer cnpj;
+    private final Long cnpj;
     private final String name;
     private final String email;
     private final String phone;
     private final Address address;
     private Set<Product> products;
 
-    public Drugstore(Integer cnpj, String name, String email, String phone, Address address) {
+    public Drugstore(Long cnpj, String name, String email, String phone, Address address) {
         this.cnpj = cnpj;
         this.name = name;
         this.email = email;
@@ -23,7 +23,7 @@ public class Drugstore implements DrugstoreEntityInterface {
         this.address = address;
     }
 
-    public Drugstore(Integer cnpj, String name, String email, String phone, Address address, Set<Product> products) {
+    public Drugstore(Long cnpj, String name, String email, String phone, Address address, Set<Product> products) {
         this.cnpj = cnpj;
         this.name = name;
         this.email = email;
@@ -32,7 +32,7 @@ public class Drugstore implements DrugstoreEntityInterface {
         this.products = products;
     }
 
-    public Integer getCnpj() {
+    public Long getCnpj() {
         return cnpj;
     }
